requirements.txt
asgiref==3.3.1
Django==3.1.5
djangorestframework==3.12.2
lorem==0.1.1
pytz==2020.5
sqlparse==0.4.1
Pillow==9.0